fe product review p2

+++ app/Services/ProductReviewService.php
@@ -2,9 +2,10 @@
 
 namespace App\Services;
 
-use App\Enum\ProductReviewStatusEnum;
+use App\Exceptions\BusinessLogicException;
 use App\Repositories\Contracts\ProductReviewRepositoryContract;
 use App\Services\BaseService;
+use Illuminate\Support\Arr;
 
 class ProductReviewService extends BaseService
 {
@@ -18,8 +19,35 @@ class ProductReviewService extends BaseService
     public function searchByAdmin($data = [])
     {
         $result = $this->productReviewRepository
-            ->with(['createdBy', 'updatedBy'])
-            ->whereColumnsLike($data['query'] ?? null, ['name'])
+            ->with(['createdBy', 'updatedBy', 'product'])
+            ->scopeQuery(function($q) use ($data) {
+                $statuses = Arr::wrap(data_get($data, 'status', []));
+
+                if (! empty($statuses)) {
+                    $q->whereIn('status', $statuses);
+                }
+
+                $ratingTypes = Arr::wrap(data_get($data, 'rating_type', []));
+
+                if (! empty($ratingTypes)) {
+                    $q->whereIn('rating_type', $ratingTypes);
+                }
+
+                $isRealUser = data_get($data, 'is_real_user', 0);
+
+                if (boolean($isRealUser)) {
+                    $q->where('is_real_user', 1);
+                }
+
+                $productId = data_get($data, 'product_id');
+
+                if (! empty($productId)) {
+                    $q->whereHas('product', function($q) use ($productId) {
+                        $q->where('id', $productId);
+                    });
+                }
+            })
+            ->whereColumnsLike($data['query'] ?? null, ['user_name', 'user_phone', 'user_email'])
             ->search([]);
 
         return $result;
@@ -27,8 +55,6 @@ class ProductReviewService extends BaseService
 
     public function create($attributes = [])
     {
-        $attributes['status'] = ProductReviewStatusEnum::PENDING;
-
         return $this->productReviewRepository->create($attributes);
     }
 
@@ -44,6 +70,12 @@ class ProductReviewService extends BaseService
 
     public function delete($id)
     {
+        $productReview = $this->show($id);
+
+        if ($productReview->is_real_user) {
+            throw new BusinessLogicException('Your registration request is pending.');
+        }
+
         return $this->productReviewRepository->delete($id);
     }
